feat: moved isISIN to validation
