--- library/src/utils/isISIN/index.ts
@@ -1,1 +1,0 @@
-export * from './isISIN.ts';
